fault-car: Clean motor_control.c
