+++ platform/stm32f3_sensors/cmds/motor_control.c
@@ -8,38 +8,16 @@
 
 #include <stdio.h>
 #include <unistd.h>
+
 #include <stm32f3xx.h>
 #include <stm32f3_discovery.h>
-#include <stm32f3xx_hal_gpio.h>
-#include <stm32f3xx_hal_rcc.h>
 
 #include <libsensors/acc.h>
 #include <libsensors/gyro.h>
 #include <libactuators/motor.h>
 
 
-static void init_pins(GPIO_TypeDef  *GPIOx, uint16_t pins) {
-	GPIO_InitTypeDef  GPIO_InitStruct;
-
-	GPIO_InitStruct.Pin = pins;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_PULLUP;
-	GPIO_InitStruct.Speed = GPIO_SPEED_HIGH;
-
-	HAL_GPIO_Init(GPIOx, &GPIO_InitStruct);
-}
-
 // --- fault detection code START
-
-static float ema = 0;
-
-#define FD_WIN_SIZE   200
-#define FD_THRESHOLD  2600
-static float f_dv[FD_WIN_SIZE] = {0};
-static int index_l = 0;
-static int index_h = FD_WIN_SIZE / 2;
-static int index = 0;
-static float diff = 0;
 
 #if 0
 #define INC_WINDOW_SIZE 50
@@ -62,12 +40,25 @@ static inline float exp_moving_average(float inc) {
 }
 #endif
 
+static float ema = 0;
+
+#define FD_WIN_SIZE   200
+#define FD_THRESHOLD  2600
+
+
 static inline float filtered_derivative(float inc) {
+	static float f_dv[FD_WIN_SIZE] = {0};
+	static int index_l = 0;
+	static int index_h = FD_WIN_SIZE / 2;
+	static int index = 0;
+	static float diff = 0;
+
 	diff += ((inc - 2 * f_dv[index_h] + f_dv[index_l]) / FD_WIN_SIZE);
 	f_dv[index] = inc;
 	index = (index + 1) % FD_WIN_SIZE;
 	index_h = (index_h + 1) % FD_WIN_SIZE;
 	index_l = (index_l + 1) % FD_WIN_SIZE;
+
 	return diff;
 }
 
@@ -98,20 +89,14 @@ static int fault_detect(void) {
 int main(int argc, char *argv[]) {
 	struct motor motor1;
 
-	BSP_LED_Init(LED4);
 	BSP_LED_Init(LED3);
 	BSP_PB_Init(BUTTON_USER, BUTTON_MODE_GPIO);
 
-	init_pins(GPIOD, MOTOR_ENABLE1 | MOTOR_INPUT1 | MOTOR_INPUT2 |
-			MOTOR_ENABLE2 | MOTOR_INPUT3 | MOTOR_INPUT4);
-
-	motor_init(&motor1, GPIOD, MOTOR_ENABLE1, MOTOR_INPUT1,
-		MOTOR_INPUT2, 0x3);
+	motor_init(&motor1, GPIOD, MOTOR_ENABLE1, MOTOR_INPUT1, MOTOR_INPUT2);
 
 	motor_enable(&motor1);
 
-	while(BSP_PB_GetState(BUTTON_USER) != SET)
-		;
+	while(BSP_PB_GetState(BUTTON_USER) != SET) ;
 
 	motor_run(&motor1, MOTOR_RUN_RIGHT);
 	{
