+++ platform/stm32f3_sensors/libactuators/motor.h
@@ -24,8 +24,6 @@ struct motor {
 	GPIO_TypeDef  *GPIOx;
 	uint16_t enable;
 	uint16_t input[2];
-	/* 10, 01, 11 */
-	uint8_t enabled_inputs:2;
 };
 
 enum motor_run_direction {
@@ -33,8 +31,8 @@ enum motor_run_direction {
 	MOTOR_RUN_RIGHT
 };
 
-extern void motor_init(struct motor *m, GPIO_TypeDef  *GPIOx,
-		uint16_t enable, uint16_t in1, uint16_t in2, uint8_t mask);
+extern void motor_init(struct motor *m, GPIO_TypeDef  *GPIOx, uint16_t enable,
+		uint16_t in1, uint16_t in2);
 
 extern void motor_enable(struct motor *m);
